+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for [snapshots](https://kit.svelte.dev/docs/snapshots) with the `capture` and `restore` functions, returned from `superForm`.
+- Added `options.selectErrorText` for selecting all text in the first invalid text field, instead of just focusing on it and the cursor is placed at the end of the text.
 
 ## [0.6.1] - 2023-03-20