+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `options.selectErrorText` for selecting all text in the first invalid text field, instead of just focusing on it and the cursor is placed at the end of the text.
 - `use:enhance` can now take all events as a parameter, in case you want to pass it on to other components.
 
+### Fixed
+
+- `superValidate` didn't use async validation, on Node! How embarrassing, but fixed now.
+
 ## [0.6.1] - 2023-03-20
 
 ### Fixed